--- a/test/kennel/models/synthetic_test_test.rb
+++ b/test/kennel/models/synthetic_test_test.rb
@@ -42,27 +42,27 @@ describe Kennel::Models::SyntheticTest do
     }
   end
 
-  describe "#as_json" do
+  describe "#working_json" do
     it "builds" do
-      assert_json_equal synthetic.as_json, expected_json
+      assert_json_equal synthetic.working_json, expected_json
     end
 
     it "caches" do
       s = synthetic
       s.expects(:locations)
-      2.times { s.as_json }
+      2.times { s.working_json }
     end
 
     it "can add id" do
-      synthetic(id: -> { 123 }).as_json[:id].must_equal 123
+      synthetic(id: -> { 123 }).working_json[:id].must_equal 123
     end
 
     it "can add all locations" do
-      synthetic(locations: -> { :all }).as_json[:locations].size.must_be :>, 5
+      synthetic(locations: -> { :all }).working_json[:locations].size.must_be :>, 5
     end
 
     it "can use super" do
-      synthetic(message: -> { super() }).as_json[:message].must_equal "\n\n@slack-foo"
+      synthetic(message: -> { super() }).working_json[:message].must_equal "\n\n@slack-foo"
     end
   end