--- a/test/kennel_test.rb
+++ b/test/kennel_test.rb
@@ -80,6 +80,37 @@ describe Kennel do
         use a different `kennel_id` when defining multiple projects/monitors/dashboards to avoid this conflict
       ERROR
     end
+
+    it "resets the working_json" do
+      project = Kennel::Models::Project.new(
+        team: TestTeam.new,
+        kennel_id: 'a_project',
+        parts: [],
+      ).instance_eval do
+        def parts
+          @parts ||= [
+            Kennel::Models::Monitor.new(
+              self,
+              type: -> { "query alert" },
+              kennel_id: -> { 'foo' },
+              query: -> { "avg(last_5m) > \#{critical}" },
+              critical: -> { 1 }
+            )
+          ]
+        end
+
+        self
+      end
+
+      Kennel::ProjectsProvider.stubs(:new).returns(OpenStruct.new(projects: [project]))
+
+      Kennel::Engine.new.generate
+      project.parts[0].working_json[:new_key] = 'whatever'
+      project.parts[0].working_json[:new_key].must_equal 'whatever'
+
+      Kennel::Engine.new.generate
+      project.parts[0].working_json[:new_key].must_be_nil
+    end
   end
 
   describe ".plan" do