… test/kennel/models/monitor_test.rb
… test/kennel/syncer_test.rb