A solution for the "static parts" problem.

 * `as_json` is now (normally) called exactly once per part for each run (therefore there's no need to memoize this method)
 * `working_json` is a kennel-internal method; it returns the memoized "working copy" of the json, defaulting to `as_json`. "Working copy" refers to the fact that the json mutates during processing.
 * `working_json!` can be used to reset the memo of `working_json`.

This means that even if a project returns a "static" part, `Kennel::Engine.new.plan` can be called more than once.