@grosser 2022-10-05
how is this better than doing `parts.each { |p| p.instance_variable_set :@json, nil }` (or a nice method that wraps this) ?