@zdrve 2022-10-05
`working_json!` is exactly the nice wrapper method which you suggest. The rest of the PR is:

 * introducing "working_json" as a name to highlight the mutating nature of the data therein
 * adding a deep clone, so as to provide stronger guarantees about the data we're working with, to make the code easier to reason about