@grosser 2022-10-05
so benchmarking shows something strange:
master:
```
Generating ... 7.96s
Storing ... 2.11s
run  10: 11.02
```

this branch:
```
Generating ... 8.55s
Storing ... 1.01s
run  10: 10.31
```

... somehow it made generation slower but storing faster ?
the thaw would be an obvious culprit for the slowness, but why is storage faster ?