@grosser 2022-10-05
looks like we can use Marshal instead:
```
>> Benchmark.realtime { 1000.times { Marshal.load Marshal.dump(a: [{b: {c: 3}}]) } }
=> 0.004929000046104193
>> Benchmark.realtime { 1000.times { deep_dup_thaw a: [{b: {c: 3}}] } }
=> 0.006802999880164862
```