@zdrve 2022-10-05
> ... for performance we can turn the duping on only when doing multiple passes ?

Let's not; sounds like a very poor complexity / payoff ratio.

All of your other feedback is 👍