@zdrve 2022-10-05
>  the thaw would be an obvious culprit for the slowness, but why is storage faster ?

This reminds me, there was something odd I wanted to investigate. I don't have an answer for your question but thank you for the unintended prompt :-)