@grosser 2022-10-05
if we simplify it down to Marshal then I'd suggest moving it into `deep_dup` in utils.rb
https://github.com/grosser/kennel/blob/43d5e1c2d4df68a80b71025ea5ab165472e61f53/test/test_helper.rb#L88-L90